+++ great_expectations/datasource/data_connector/util.py
@@ -320,6 +320,26 @@ def sanitize_prefix(text: str) -> str:
     return os.path.join(text, "")
 
 
+def sanitize_prefix_for_s3(text: str) -> str:
+    """
+    Takes in a given user-prefix and cleans it to work with file-system traversal methods
+    (i.e. add '/' to the end of a string meant to represent a directory)
+
+    Customized for S3 paths, ignoring the path separator used by the host OS
+    """
+    text = text.strip()
+    if not text:
+        return text
+
+    path_parts = text.split("/")
+    if not path_parts:  # Empty prefix
+        return text
+    elif "." in path_parts[-1]:  # File, not folder
+        return text
+    else:  # Folder, should have trailing /
+        return f"{text.rstrip('/')}/"
+
+
 def normalize_directory_path(
     dir_path: str, root_directory_path: Optional[str] = None
 ) -> str:
