+++ tests/datasource/data_connector/test_configured_asset_s3_data_connector.py
@@ -18,9 +18,10 @@ from great_expectations.core.batch import (
     IDDict,
 )
 from great_expectations.data_context.util import instantiate_class_from_config
-from great_expectations.datasource.data_connector import (
-    ConfiguredAssetS3DataConnector,
-    FilePathDataConnector,
+from great_expectations.datasource.data_connector import ConfiguredAssetS3DataConnector
+from great_expectations.datasource.data_connector.util import (
+    sanitize_prefix,
+    sanitize_prefix_for_s3,
 )
 from great_expectations.execution_engine import PandasExecutionEngine
 
@@ -1125,8 +1126,8 @@ assets:
 
 def test_sanitize_prefix_behaves_the_same_as_local_files():
     def check_sameness(prefix, expected_output):
-        s3_sanitized = ConfiguredAssetS3DataConnector.sanitize_prefix_for_s3(prefix)
-        file_system_sanitized = FilePathDataConnector.sanitize_prefix(prefix)
+        s3_sanitized = sanitize_prefix_for_s3(prefix)
+        file_system_sanitized = sanitize_prefix(prefix)
         if os.sep == "\\":  # Fix to ensure tests work on Windows
             file_system_sanitized = file_system_sanitized.replace("\\", "/")
